Split authorities-noting-inherent into 2 crates
Avoids runtime dependency on libp2p
