+++ Cargo.toml
@@ -143,6 +143,7 @@ westend-emulated-chain = { path = "chains/xcm-integration-tests/emulated/chains/
 westend-system-emulated-network = { path = "chains/xcm-integration-tests/emulated/networks/westend-system" }
 
 tp-author-noting-inherent = { path = "primitives/author-noting-inherent", default-features = false }
+tp-author-noting-inherent-core = { path = "primitives/author-noting-inherent-core", default-features = false }
 tp-bridge = { path = "primitives/bridge", default-features = false }
 tp-data-preservers-common = { path = "primitives/data-preservers-common", default-features = false }
 tp-invulnerables-filter-common = { path = "primitives/invulnerables-filter-common", default-features = false }
@@ -154,23 +155,23 @@ tp-xcm-commons = { path = "primitives/xcm-commons", default-features = false }
 tp-xcm-core-buyer = { path = "primitives/xcm-core-buyer", default-features = false }
 
 # Dancekit (wasm)
-ccp-authorities-noting-inherent = { git = "https://github.com/moondance-labs/dancekit", branch = "tanssi-polkadot-stable2503", default-features = false }
-ccp-xcm = { git = "https://github.com/moondance-labs/dancekit", branch = "tanssi-polkadot-stable2503", default-features = false }
-dp-chain-state-snapshot = { git = "https://github.com/moondance-labs/dancekit", branch = "tanssi-polkadot-stable2503", default-features = false }
-dp-collator-assignment = { git = "https://github.com/moondance-labs/dancekit", branch = "tanssi-polkadot-stable2503", default-features = false }
-dp-consensus = { git = "https://github.com/moondance-labs/dancekit", branch = "tanssi-polkadot-stable2503", default-features = false }
-dp-impl-tanssi-pallets-config = { git = "https://github.com/moondance-labs/dancekit", branch = "tanssi-polkadot-stable2503", default-features = false }
-dp-slot-duration-runtime-api = { git = "https://github.com/moondance-labs/dancekit", branch = "tanssi-polkadot-stable2503", default-features = false }
+ccp-authorities-noting-inherent = { git = "https://github.com/moondance-labs/dancekit", branch = "tomasz-authorities-noting-inherent-core", default-features = false }
+ccp-xcm = { git = "https://github.com/moondance-labs/dancekit", branch = "tomasz-authorities-noting-inherent-core", default-features = false }
+dp-chain-state-snapshot = { git = "https://github.com/moondance-labs/dancekit", branch = "tomasz-authorities-noting-inherent-core", default-features = false }
+dp-collator-assignment = { git = "https://github.com/moondance-labs/dancekit", branch = "tomasz-authorities-noting-inherent-core", default-features = false }
+dp-consensus = { git = "https://github.com/moondance-labs/dancekit", branch = "tomasz-authorities-noting-inherent-core", default-features = false }
+dp-impl-tanssi-pallets-config = { git = "https://github.com/moondance-labs/dancekit", branch = "tomasz-authorities-noting-inherent-core", default-features = false }
+dp-slot-duration-runtime-api = { git = "https://github.com/moondance-labs/dancekit", branch = "tomasz-authorities-noting-inherent-core", default-features = false }
 
-dp-core = { git = "https://github.com/moondance-labs/dancekit", branch = "tanssi-polkadot-stable2503", default-features = false }
-pallet-cc-authorities-noting = { git = "https://github.com/moondance-labs/dancekit", branch = "tanssi-polkadot-stable2503", default-features = false }
-pallet-xcm-executor-utils = { git = "https://github.com/moondance-labs/dancekit", branch = "tanssi-polkadot-stable2503", default-features = false }
-test-relay-sproof-builder = { git = "https://github.com/moondance-labs/dancekit", branch = "tanssi-polkadot-stable2503", default-features = false }
+dp-core = { git = "https://github.com/moondance-labs/dancekit", branch = "tomasz-authorities-noting-inherent-core", default-features = false }
+pallet-cc-authorities-noting = { git = "https://github.com/moondance-labs/dancekit", branch = "tomasz-authorities-noting-inherent-core", default-features = false }
+pallet-xcm-executor-utils = { git = "https://github.com/moondance-labs/dancekit", branch = "tomasz-authorities-noting-inherent-core", default-features = false }
+test-relay-sproof-builder = { git = "https://github.com/moondance-labs/dancekit", branch = "tomasz-authorities-noting-inherent-core", default-features = false }
 
-dp-container-chain-genesis-data = { git = "https://github.com/moondance-labs/dancekit", branch = "tanssi-polkadot-stable2503", default-features = false }
+dp-container-chain-genesis-data = { git = "https://github.com/moondance-labs/dancekit", branch = "tomasz-authorities-noting-inherent-core", default-features = false }
 
 # Dancekit (client)
-dc-orchestrator-chain-interface = { git = "https://github.com/moondance-labs/dancekit", branch = "tanssi-polkadot-stable2503" }
+dc-orchestrator-chain-interface = { git = "https://github.com/moondance-labs/dancekit", branch = "tomasz-authorities-noting-inherent-core" }
 
 # Moonkit (wasm)
 async-backing-primitives = { git = "https://github.com/moondance-labs/moonkit", branch = "tanssi-polkadot-stable2503", default-features = false }
