+++ primitives/author-noting-inherent-core/src/client_side.rs
@@ -1,0 +1,20 @@
+use crate::OwnParachainInherentData;
+
+// Implementation of InherentDataProvider
+#[async_trait::async_trait]
+impl sp_inherents::InherentDataProvider for OwnParachainInherentData {
+    async fn provide_inherent_data(
+        &self,
+        inherent_data: &mut sp_inherents::InherentData,
+    ) -> Result<(), sp_inherents::Error> {
+        inherent_data.put_data(crate::INHERENT_IDENTIFIER, &self)
+    }
+
+    async fn try_handle_error(
+        &self,
+        _: &sp_inherents::InherentIdentifier,
+        _: &[u8],
+    ) -> Option<Result<(), sp_inherents::Error>> {
+        None
+    }
+}
