+++ primitives/author-noting-inherent-core/src/lib.rs
@@ -14,33 +14,34 @@
 // You should have received a copy of the GNU General Public License
 // along with Tanssi.  If not, see <http://www.gnu.org/licenses/>
 
-//! # Author Noting Inherent Primitives
+//! # Author Noting Inherent Core Primitives
 //!
-//! This crate defines the client-side primitives that should be taken into account when building
-//! the author-noting pallet inherent.
+//! This crate defines the runtime primitives for the author-noting pallet inherent.
 //!
-//! Client-side methods should be implemented in `tp-author-noting-inherent-core` crate instead.
+//! Client-side methods should be implemented in `tp-author-noting-inherent` crate instead.
 //!
 //! In particular, this crate contains:
-//! - The client side trait implementations to introduce the inherent
-//! - The mock version that gets used both in test files and manual seal
-//! - The sproof builder that generates a fake proof that mimics the relay chain sproof
-//! - Tests
+//! - The Inherent identifier
+//! - The `OwnParachainInherentData` struct
+//! - The `InherentDataProvider` trait impl
 
 #![cfg_attr(not(feature = "std"), no_std)]
 
+use {
+    parity_scale_codec::{Decode, DecodeWithMemTracking, Encode},
+    scale_info::TypeInfo,
+    sp_inherents::InherentIdentifier,
+};
+
 #[cfg(feature = "std")]
 mod client_side;
-#[cfg(feature = "std")]
-pub use client_side::*;
 
-#[cfg(feature = "std")]
-mod mock;
-#[cfg(feature = "std")]
-pub use mock::*;
+#[derive(
+    Encode, Decode, DecodeWithMemTracking, sp_core::RuntimeDebug, Clone, PartialEq, TypeInfo,
+)]
+pub struct OwnParachainInherentData {
+    pub relay_storage_proof: sp_trie::StorageProof,
+}
 
-#[cfg(test)]
-mod tests;
-
-pub use tp_author_noting_inherent_core::OwnParachainInherentData;
-pub use tp_author_noting_inherent_core::INHERENT_IDENTIFIER;
+// Identifier of the author-noting inherent
+pub const INHERENT_IDENTIFIER: InherentIdentifier = *b"auno1337";
